requirements.txt

streamlit
transformers
torch
python-docx
git+https://github.com/openai/whisper.git